+++ src/open_global_liquidity/__init__.py
@@ -1,3 +1,3 @@
 """Open Global Liquidity public-data research package."""
 
-__version__ = "0.3.0"
+__version__ = "0.4.0.dev0"